% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch_error_as_na.R
\name{try_catch_error_as_na}
\alias{try_catch_error_as_na}
\title{Wrapper that returns NA if evaluating expression returns an error}
\usage{
try_catch_error_as_na(expr)
}
\arguments{
\item{expr}{expression}
}
\description{
Wrapper that returns NA if evaluating expression returns an error
}
